+++ packages/adapter-cloudflare-workers/files/entry.js
@@ -41,7 +41,7 @@ async function handle(event) {
 		if (rendered) {
 			return new Response(rendered.body, {
 				status: rendered.status,
-				headers: rendered.headers
+				headers: makeHeaders(rendered.headers)
 			});
 		}
 	} catch (e) {
@@ -57,4 +57,23 @@ async function handle(event) {
 /** @param {Request} request */
 async function read(request) {
 	return new Uint8Array(await request.arrayBuffer());
+}
+
+/**
+ * @param {Record<string, string | string[]>} headers
+ * @returns {Request}
+ */
+function makeHeaders(headers) {
+	const result = new Headers();
+	for (const header in headers) {
+		const value = headers[header];
+		if (typeof value === 'string') {
+			result.set(header, value);
+			continue;
+		}
+		for (const sub of value) {
+			result.append(header, sub);
+		}
+	}
+	return result;
 }