+++ packages/kit/src/cli.js
@@ -4,6 +4,7 @@ import colors from 'kleur';
 import * as ports from 'port-authority';
 import { load_config } from './core/config/index.js';
 import { networkInterfaces, release } from 'os';
+import { coalesce_to_error, has_error_code } from './utils/error.js';
 
 async function get_config() {
 	// TODO this is temporary, for the benefit of early adopters
@@ -23,11 +24,12 @@ async function get_config() {
 
 	try {
 		return await load_config();
-	} catch (error) {
+	} catch (err) {
+		const error = coalesce_to_error(err);
 		let message = error.message;
 
 		if (
-			error.code === 'MODULE_NOT_FOUND' &&
+			has_error_code(error, 'MODULE_NOT_FOUND') &&
 			/Cannot find module svelte\.config\./.test(error.message)
 		) {
 			message = 'Missing svelte.config.js';
@@ -36,16 +38,19 @@ async function get_config() {
 		}
 
 		console.error(colors.bold().red(message));
-		console.error(colors.grey(error.stack));
+		if (error.stack) {
+			console.error(colors.grey(error.stack));
+		}
 		process.exit(1);
 	}
 }
 
-/** @param {Error} error */
+/** @param {unknown} error */
 function handle_error(error) {
-	console.log(colors.bold().red(`> ${error.message}`));
-	if (error.stack) {
-		console.log(colors.gray(error.stack));
+	const err = coalesce_to_error(error);
+	console.log(colors.bold().red(`> ${err.message}`));
+	if (err.stack) {
+		console.log(colors.gray(err.stack));
 	}
 	process.exit(1);
 }