+++ packages/kit/src/packaging/test/fixtures/exports-replace/svelte.config.js
@@ -1,0 +1,12 @@
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		package: {
+			exports: {
+				exclude: ['**']
+			}
+		}
+	}
+};
+
+export default config;