+++ packages/kit/test/apps/basics/src/routes/redirect/_tests.js
@@ -5,7 +5,7 @@ export default function (test, is_dev) {
 	test('redirect', '/redirect', async ({ base, page, clicknav }) => {
 		await clicknav('[href="/redirect/a"]');
 
-		assert.equal(await page.url(), `${base}/redirect/c`);
+		assert.equal(page.url(), `${base}/redirect/c`);
 		assert.equal(await page.textContent('h1'), 'c');
 	});
 
@@ -14,7 +14,7 @@ export default function (test, is_dev) {
 
 		if (js) {
 			await page.waitForTimeout(100);
-			assert.equal(await page.url(), `${base}/redirect/loopy/a`);
+			assert.equal(page.url(), `${base}/redirect/loopy/a`);
 			assert.equal(await page.textContent('h1'), '500');
 			assert.equal(
 				await page.textContent('#message'),
@@ -22,14 +22,14 @@ export default function (test, is_dev) {
 			);
 		} else {
 			// there's not a lot we can do to handle server-side redirect loops
-			assert.equal(await page.url(), 'chrome-error://chromewebdata/');
+			assert.equal(page.url(), 'chrome-error://chromewebdata/');
 		}
 	});
 
 	test('errors on missing status', '/redirect', async ({ base, page, clicknav }) => {
 		await clicknav('[href="/redirect/missing-status/a"]');
 
-		assert.equal(await page.url(), `${base}/redirect/missing-status/a`);
+		assert.equal(page.url(), `${base}/redirect/missing-status/a`);
 		assert.equal(await page.textContent('h1'), '500');
 		assert.equal(
 			await page.textContent('#message'),
@@ -40,7 +40,7 @@ export default function (test, is_dev) {
 	test('errors on invalid status', '/redirect', async ({ base, page, clicknav }) => {
 		await clicknav('[href="/redirect/missing-status/b"]');
 
-		assert.equal(await page.url(), `${base}/redirect/missing-status/b`);
+		assert.equal(page.url(), `${base}/redirect/missing-status/b`);
 		assert.equal(await page.textContent('h1'), '500');
 		assert.equal(
 			await page.textContent('#message'),